mvp_backend/parser/requirements.txt
python-docx>=1.1.2
pdfplumber>=0.11.4
openpyxl>=3.1.5
